# docs/openapi.json
{
  "components": {
    "schemas": {
      "DirectoryListingResponse": {
        "description": "Response body for server-side directory browsing.",
        "properties": {
          "directories": {
            "items": {
              "type": "string"
            },
            "title": "Directories",
            "type": "array"
          },
          "parent_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Path"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "parent_path",
          "directories"
        ],
        "title": "DirectoryListingResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PurgeRequest": {
        "description": "Request body for orphan metadata cleanup.",
        "properties": {
          "retention_days": {
            "default": 3,
            "minimum": 0.0,
            "title": "Retention Days",
            "type": "integer"
          }
        },
        "title": "PurgeRequest",
        "type": "object"
      },
      "RootCreate": {
        "description": "Request body for creating an indexed root.",
        "properties": {
          "path": {
            "format": "path",
            "title": "Path",
            "type": "string"
          },
          "recursive": {
            "default": true,
            "title": "Recursive",
            "type": "boolean"
          },
          "replace_covered_roots": {
            "default": false,
            "title": "Replace Covered Roots",
            "type": "boolean"
          }
        },
        "required": [
          "path"
        ],
        "title": "RootCreate",
        "type": "object"
      },
...
      "TagCreate": {
        "description": "Request body for creating or applying a tag.",
        "properties": {
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "TagCreate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Phag API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/filesystem/directories": {
      "get": {
        "description": "List server-local directories for root selection.",
        "operationId": "get_directories_filesystem_directories_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "path",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryListingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Directories"
      }
    },
...
    "/images": {
      "get": {
        "description": "List active library images.",
        "operationId": "get_images_images_get",
        "parameters": [
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Tag",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "root_id",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "integer"
              },
              "title": "Root Id",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "include_untagged",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Untagged",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "tag_match",
            "required": false,
            "schema": {
              "default": "and",
              "title": "Tag Match",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "filename",
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_direction",
            "required": false,
            "schema": {
              "default": "asc",
              "title": "Sort Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Images Images Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Images"
      }
    },
    "/images/{image_id}/file": {
      "get": {
        "description": "Serve one active source file for an indexed image.",
        "operationId": "get_image_file_images__image_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Image File"
      }
    },
    "/images/{image_id}/tags": {
      "get": {
        "description": "List tags for one image.",
        "operationId": "get_image_tags_images__image_id__tags_get",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Image Tags Images  Image Id  Tags Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Image Tags"
      },
      "post": {
        "description": "Apply a tag to one image, creating the tag if needed.",
        "operationId": "tag_image_images__image_id__tags_post",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Tag Image Images  Image Id  Tags Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tag Image"
      }
    },
    "/images/{image_id}/tags/{tag_id}": {
      "delete": {
        "description": "Remove one tag association from an image.",
        "operationId": "untag_image_images__image_id__tags__tag_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "tag_id",
            "required": true,
            "schema": {
              "title": "Tag Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Untag Image Images  Image Id  Tags  Tag Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Untag Image"
      }
    },
    "/orphans/purge": {
      "post": {
        "description": "Purge expired orphan metadata.",
        "operationId": "purge_orphan_metadata_orphans_purge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurgeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Purge Orphan Metadata Orphans Purge Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Purge Orphan Metadata"
      }
    },
    "/roots": {
      "get": {
        "description": "List configured indexed roots.",
        "operationId": "get_roots_roots_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_disabled",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Disabled",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Roots Roots Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Roots"
      },
      "post": {
        "description": "Register a folder as an indexed root.",
        "operationId": "create_root_roots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RootCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Root Roots Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Root"
      }
    },
    "/roots/{root_id}": {
      "delete": {
        "description": "Disable an indexed root.",
        "operationId": "delete_root_roots__root_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "root_id",
            "required": true,
            "schema": {
              "title": "Root Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Delete Root Roots  Root Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Root"
      },
      "patch": {
        "description": "Update scan options for an indexed root.",
        "operationId": "patch_root_roots__root_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "root_id",
            "required": true,
            "schema": {
              "title": "Root Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RootUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Patch Root Roots  Root Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Root"
      }
    },
...
    "/scans": {
      "post": {
        "description": "Run an indexing scan immediately.",
        "operationId": "create_scan_scans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Create Scan Scans Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scan"
      }
    },
...
  }
}